+++ src/mesh/Mesh.cpp
@@ -122,14 +122,11 @@ void Mesh<VertexPositionType>::addPolygonToMesh(const Polygon& polygon) {
   }
   CHECK_EQ(vtx_ids.size(), polygon_dimension_);
 
-
   // DO NOT sort vtx_ids, or the normals will flip!
   VertexIds sorted_vtx_ids = vtx_ids;
   std::sort(sorted_vtx_ids.begin(), sorted_vtx_ids.end());
-  const auto& face_hash =
-      UtilsNumerical::hashTriplet(sorted_vtx_ids[0],
-                                  sorted_vtx_ids[1],
-                                  sorted_vtx_ids[2]);
+  const auto& face_hash = UtilsNumerical::hashTriplet(
+      sorted_vtx_ids[0], sorted_vtx_ids[1], sorted_vtx_ids[2]);
   const auto& it = face_hashes_.find(face_hash);
 
   // Check the triangle is not already in the mesh
@@ -142,9 +139,15 @@ void Mesh<VertexPositionType>::addPolygonToMesh(const Polygon& polygon) {
       // Triangle already exists!
       triangle_in_mesh = true;
       CHECK(it->second);
-      CHECK_EQ(adjacency_matrix_.at<uint8_t>(sorted_vtx_ids[0], sorted_vtx_ids[1]), 1u);
-      CHECK_EQ(adjacency_matrix_.at<uint8_t>(sorted_vtx_ids[1], sorted_vtx_ids[2]), 1u);
-      CHECK_EQ(adjacency_matrix_.at<uint8_t>(sorted_vtx_ids[2], sorted_vtx_ids[0]), 1u);
+      CHECK_EQ(
+          adjacency_matrix_.at<uint8_t>(sorted_vtx_ids[0], sorted_vtx_ids[1]),
+          1u);
+      CHECK_EQ(
+          adjacency_matrix_.at<uint8_t>(sorted_vtx_ids[1], sorted_vtx_ids[2]),
+          1u);
+      CHECK_EQ(
+          adjacency_matrix_.at<uint8_t>(sorted_vtx_ids[2], sorted_vtx_ids[0]),
+          1u);
     } else {
       triangle_in_mesh = false;
     }
@@ -163,7 +166,6 @@ void Mesh<VertexPositionType>::addPolygonToMesh(const Polygon& polygon) {
       polygons_mesh_.push_back(static_cast<int>(vtx_id));
     }
 
-
     // Update adjacency matrix
     if (!triangle_maybe_already_in_mesh) {
       // There are new vertices!
@@ -173,7 +175,7 @@ void Mesh<VertexPositionType>::addPolygonToMesh(const Polygon& polygon) {
       VertexIds sorted_vtx_ids = vtx_ids;
       std::sort(sorted_vtx_ids.begin(), sorted_vtx_ids.end());
       for (const auto& sorted_vtx_id : sorted_vtx_ids) {
-        if (sorted_vtx_id >= adjacency_matrix_.rows) {
+        if (sorted_vtx_id >= static_cast<size_t>(adjacency_matrix_.rows)) {
           // Non-existing vertex! Add row/col. Careful! vtx_ids ordering
           // matters! First order them!
           CHECK(adjacency_matrix_.rows != 0 && adjacency_matrix_.cols != 0);
@@ -270,7 +272,8 @@ bool Mesh<VertexPositionType>::getPolygon(const size_t& polygon_idx,
     return false;
   };
 
-  bool has_normals = vertices_mesh_.rows == vertices_mesh_normal_.size();
+  bool has_normals =
+      static_cast<size_t>(vertices_mesh_.rows) == vertices_mesh_normal_.size();
   CHECK_EQ(vertices_mesh_.rows, vertices_mesh_color_.rows);
   size_t idx_in_polygon_mesh = polygon_idx * (polygon_dimension_ + 1);
   polygon->resize(polygon_dimension_);
@@ -278,13 +281,13 @@ bool Mesh<VertexPositionType>::getPolygon(const size_t& polygon_idx,
     const int32_t& row_id_pt_j =
         polygons_mesh_.at<int32_t>(idx_in_polygon_mesh + j + 1);
     CHECK(vertex_to_lmk_id_map_.find(row_id_pt_j) !=
-           vertex_to_lmk_id_map_.end());
+          vertex_to_lmk_id_map_.end());
     CHECK_LT(row_id_pt_j, vertices_mesh_.rows);
     polygon->at(j) = Vertex<VertexPositionType>(
         vertex_to_lmk_id_map_.at(row_id_pt_j),
         vertices_mesh_.at<VertexPositionType>(row_id_pt_j),
         vertices_mesh_color_.at<VertexColorRGB>(row_id_pt_j),
-        has_normals? vertices_mesh_normal_.at(row_id_pt_j) : VertexNormal());
+        has_normals ? vertices_mesh_normal_.at(row_id_pt_j) : VertexNormal());
   }
   return true;
 }
@@ -379,11 +382,11 @@ void Mesh<VertexPositionType>::computePerVertexNormals() {
         (counts.at(p1_idx) * vertices_mesh_normal_.at(p1_idx) + normal) /
         (counts.at(p1_idx) + 1.0);
     vertices_mesh_normal_.at(p2_idx) =
-        counts.at(p2_idx) * vertices_mesh_normal_.at(p2_idx) + normal /
-        (counts.at(p2_idx) + 1.0);
+        counts.at(p2_idx) * vertices_mesh_normal_.at(p2_idx) +
+        normal / (counts.at(p2_idx) + 1.0);
     vertices_mesh_normal_.at(p3_idx) =
-        counts.at(p3_idx) * vertices_mesh_normal_.at(p3_idx) + normal /
-        (counts.at(p3_idx) + 1.0);
+        counts.at(p3_idx) * vertices_mesh_normal_.at(p3_idx) +
+        normal / (counts.at(p3_idx) + 1.0);
     // assumes non-zero normals...
     vertices_mesh_normal_.at(p1_idx) /=
         cv::norm(vertices_mesh_normal_.at(p1_idx));
@@ -471,9 +474,8 @@ void Mesh<VertexPositionType>::getPolygonsMeshToMat(
 }
 
 template <typename VertexPositionType>
-cv::Mat Mesh<VertexPositionType>::getColorsMesh(
-    const bool& safe) const {
-  return safe? vertices_mesh_color_.clone() : vertices_mesh_color_;
+cv::Mat Mesh<VertexPositionType>::getColorsMesh(const bool& safe) const {
+  return safe ? vertices_mesh_color_.clone() : vertices_mesh_color_;
 }
 
 template <typename VertexPositionType>
